+++ CHANGELOG.md
@@ -41,6 +41,7 @@ See the breaking changes of the releases 0.4.0 and 0.3.0 please.
 ### Changed
 
 - [package]: Updated dependencies.
+- [update]: The attributes update is skipped when state is off only for the domains light, switch and fan.
 
 <a href="https://www.buymeacoffee.com/luligugithub">
   <img src="bmc-button.svg" alt="Buy me a coffee" width="80">